// packages/dataapi/src/types/WalletSaved.ts
export type WalletSaved = {
  count: 2;
  wallets: [
    {
      uid: "u_9f2c";
      wallet: "9BB6NFEcjBCtnNLFko2FqVQBq8HHM13kCyYcdQbgpump";
      note: "";
      created_at: 1787891901;
    },
    {
      uid: "u_9f2c";
      wallet: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU";
      note: "whale - exited";
      created_at: 1787891892;
    },
  ];
};